+++ src/main/java/com/chatspellcheck/ChatSpellcheckPlugin.java
@@ -5,7 +5,6 @@ import java.util.concurrent.ScheduledExecutorService;
 import javax.inject.Inject;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.eventbus.EventBus;
-import net.runelite.client.input.KeyManager;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 import net.runelite.client.ui.overlay.OverlayManager;
@@ -22,9 +21,6 @@ public class ChatSpellcheckPlugin extends Plugin
 
 	@Inject
 	private OverlayManager overlayManager;
-
-	@Inject
-	private KeyManager keyManager;
 
 	@Inject
 	private ScheduledExecutorService executor;
@@ -53,23 +49,23 @@ public class ChatSpellcheckPlugin extends Plugin
 		ignoreListStore.load();
 		executor.execute(dictionary::load);
 
+		eventBus.register(ignoreListStore);
 		eventBus.register(chatInputTracker);
 		eventBus.register(menuManager);
 		eventBus.register(sendGuard);
 
 		overlayManager.add(overlay);
-		keyManager.registerKeyListener(sendGuard);
 	}
 
 	@Override
 	protected void shutDown()
 	{
-		keyManager.unregisterKeyListener(sendGuard);
 		overlayManager.remove(overlay);
 
 		eventBus.unregister(sendGuard);
 		eventBus.unregister(menuManager);
 		eventBus.unregister(chatInputTracker);
+		eventBus.unregister(ignoreListStore);
 
 		chatInputTracker.reset();
 		sendGuard.reset();